backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
tensorflow==2.13.0
librosa==0.10.1
numpy==1.24.3
scipy==1.11.4
python-dotenv==1.0.0
